==> Preprocess/Script1_generatedata_w1w_June20192.Rmd <==
---
title: "Untitled"
author: "Angel Garcia de la Garza"
date: "2/17/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(haven)
library(tidyverse)


```



```{r}

## Load Data

w1mother <- haven::read_sas("/Users/angelgar/NESARC/Data/w1mother.sas7bdat")
w2mother <- read_csv("/Users/angelgar/NESARC/Data/w2mother_suicideonly.csv")
load("/Users/angelgar/NESARC/Data/w1w2suicidedata.RData")

## Generate Names

names(w1mother) <- toupper(names(w1mother))
keep.data <- toupper(names(w1mother)) %in% c(w1vars1$code, "INCPER3", "INCFAM3")

## Add Income Variables INCPER3 INCFAM3

not.mother <-  w1vars1$code[!(w1vars1$code %in% toupper(names(w1mother)))]

## Filter and Merge

w1mother.codebook <- w1mother[,keep.data]
              
w1w2suicide.v2 <- w2mother %>%
                  select(IDNUM, W2S4AQ3A16, W2S14Q17A:W2S14Q17C, W2AGE) %>%
                  left_join(w1mother.codebook, by = "IDNUM")

save("w1w2suicide.v2", 
     file = "/Users/angelgar/NESARC/Data/w1w2suicidedata_06_04_2019.RData")

complete.id <- w1w2suicide.v2$IDNUM

w1missing <- w1mother.codebook %>%
                filter(!(IDNUM %in% complete.id))

save("w1missing", 
     file = "/Users/angelgar/NESARC/Data/w1suicide_data_losstofollowup.RData")


```
