==> Preprocess/Script2_NESARC_CaseDefition_June2019.Rmd <==
---
title: "NESARC New Attempts Machine Learning 253 Sample"
author: "Angel Garcia de la Garza"
date: "2/5/2019"
output: html_document
---


```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, 
                      results = 'asis',
                      cache = TRUE,
                      message = FALSE,
                      warning = FALSE,
                      cache.lazy = FALSE)

library(pROC)
library(doParallel)
library(beepr)
library(summarytools)
library(psych)
library(knitr)
library(forcats)
library(furrr)
library(randomForest)
library(gbm)
library(e1071)
library(caret)
library(glmnet)
library(tidyverse)



```

## R Markdown


```{r load data, cache = TRUE}

#load("/Volumes/BIOS/Projects/NESARC/ANALYSIS/suicide/angelgar/Data/w1w2suicidedata_v2_2019_02_20.RData")

load("/Users/angelgar/NESARC/Data/w1w2suicidedata_06_04_2019.RData")

```


# Remove previous attempters

```{r}

w1w2suicide <- w1w2suicide.v2
rm(w1w2suicide.v2)

w1w2suicide.v3 <- w1w2suicide %>%
                mutate(suicide_sli = W2S4AQ3A16,
                       suicide_sli = ifelse(is.na(suicide_sli),0,suicide_sli),
                       suicide_sli = ifelse(suicide_sli == 2,0,suicide_sli),
                       suicide_sli = ifelse(suicide_sli == 9,1,suicide_sli)) %>%
                mutate(suicide_life = ifelse(W2S14Q17A == 1,1,0),
                       suicide_age = ifelse(W2S14Q17C == 0 &  (4 <= W2S14Q17B & W2S14Q17B <= 77),
                                            W2S14Q17B,
                                     ifelse(10 <= W2S14Q17C & W2S14Q17C <= 80,
                                            W2S14Q17C, NA))) %>%
                mutate(age_diff = W2AGE - suicide_age,
                       suicide_new = ifelse(0 <= age_diff & age_diff <= 3,
                                            1, 0),
                       suicide_new = ifelse(is.na(suicide_new), 0, suicide_new))
 

```



```{r}

save(w1w2suicide.v3, file = "/Users/angelgar/NESARC/Data/w1w2suicidedata_WithCases_06_04_2019.RData")


```
